doc(client/db): document the crate and it's public symbols
