+++ crates/client/db/src/db_opening_utils/parity_db_adapter.rs
@@ -1,0 +1,63 @@
+// SPDX-License-Identifier: GPL-3.0-or-later WITH Classpath-exception-2.0
+// This file is part of Frontier.
+//
+// Copyright (c) 2022 Parity Technologies (UK) Ltd.
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see <https://www.gnu.org/licenses/>.
+
+use sp_database::error::DatabaseError;
+use sp_database::{Change, ColumnId, Database, Transaction};
+
+fn handle_err<T>(result: parity_db::Result<T>) -> T {
+    match result {
+        Ok(r) => r,
+        Err(e) => {
+            panic!("Critical database error: {:?}", e);
+        }
+    }
+}
+
+pub struct DbAdapter(pub parity_db::Db);
+
+impl<H: Clone + AsRef<[u8]>> Database<H> for DbAdapter {
+    fn commit(&self, transaction: Transaction<H>) -> Result<(), DatabaseError> {
+        handle_err(self.0.commit(transaction.0.into_iter().map(|change| match change {
+            Change::Set(col, key, value) => (col as u8, key, Some(value)),
+            Change::Remove(col, key) => (col as u8, key, None),
+            _ => unimplemented!(),
+        })));
+
+        Ok(())
+    }
+
+    fn get(&self, col: ColumnId, key: &[u8]) -> Option<Vec<u8>> {
+        handle_err(self.0.get(col as u8, key))
+    }
+
+    fn contains(&self, col: ColumnId, key: &[u8]) -> bool {
+        handle_err(self.0.get_size(col as u8, key)).is_some()
+    }
+
+    fn value_size(&self, col: ColumnId, key: &[u8]) -> Option<usize> {
+        handle_err(self.0.get_size(col as u8, key)).map(|s| s as usize)
+    }
+
+    fn supports_ref_counting(&self) -> bool {
+        true
+    }
+
+    fn sanitize_key(&self, key: &mut Vec<u8>) {
+        let _prefix = key.drain(0..key.len() - crate::DB_HASH_LEN);
+    }
+}
