+++ crates/client/db/src/mapping_db.rs
@@ -9,13 +9,15 @@ use sp_runtime::traits::Block as BlockT;
 
 use crate::DbHash;
 
-#[derive(Clone, Debug, Eq, PartialEq, Encode, Decode)]
-pub struct TransactionMetadata<Block: BlockT> {
-    pub block_hash: Block::Hash,
-    pub starknet_block_hash: H256,
-    pub starknet_index: u32,
-}
+// TODO: use implement whe we have transactions
+// #[derive(Clone, Debug, Eq, PartialEq, Encode, Decode)]
+// pub struct TransactionMetadata<Block: BlockT> {
+//     pub block_hash: Block::Hash,
+//     pub starknet_block_hash: H256,
+//     pub starknet_index: u32,
+// }
 
+/// The mapping to write in db
 #[derive(Debug)]
 pub struct MappingCommitment<Block: BlockT> {
     pub block_hash: Block::Hash,
@@ -24,6 +26,7 @@ pub struct MappingCommitment<Block: BlockT> {
     // pub starknet_transaction_hashes: Vec<H256>,
 }
 
+/// Allow interaction with the mapping db
 pub struct MappingDb<Block: BlockT> {
     pub(crate) db: Arc<dyn Database<DbHash>>,
     pub(crate) write_lock: Arc<Mutex<()>>,
@@ -31,6 +34,7 @@ pub struct MappingDb<Block: BlockT> {
 }
 
 impl<Block: BlockT> MappingDb<Block> {
+    /// Check if the given block hash has already been processed
     pub fn is_synced(&self, block_hash: &Block::Hash) -> Result<bool, String> {
         match self.db.get(crate::columns::SYNCED_MAPPING, &block_hash.encode()) {
             Some(raw) => Ok(bool::decode(&mut &raw[..]).map_err(|e| format!("{:?}", e))?),
@@ -38,6 +42,10 @@ impl<Block: BlockT> MappingDb<Block> {
         }
     }
 
+    /// Return the hash of the Substrate block wrapping the Starknet block with given hash
+    ///
+    /// Under some circumstances it can return multiples blocks hashes, meaning that the result has
+    /// to be checked against the actual blockchain state in order to find the good one.
     pub fn block_hash(&self, starknet_block_hash: &H256) -> Result<Option<Vec<Block::Hash>>, String> {
         match self.db.get(crate::columns::BLOCK_MAPPING, &starknet_block_hash.encode()) {
             Some(raw) => Ok(Some(Vec::<Block::Hash>::decode(&mut &raw[..]).map_err(|e| format!("{:?}", e))?)),
@@ -45,16 +53,18 @@ impl<Block: BlockT> MappingDb<Block> {
         }
     }
 
-    pub fn transaction_metadata(
-        &self,
-        starknet_transaction_hash: &H256,
-    ) -> Result<Vec<TransactionMetadata<Block>>, String> {
-        match self.db.get(crate::columns::TRANSACTION_MAPPING, &starknet_transaction_hash.encode()) {
-            Some(raw) => Ok(Vec::<TransactionMetadata<Block>>::decode(&mut &raw[..]).map_err(|e| format!("{:?}", e))?),
-            None => Ok(Vec::new()),
-        }
-    }
+    // TODO: implement whe we have transactions
+    // pub fn transaction_metadata(
+    //     &self,
+    //     starknet_transaction_hash: &H256,
+    // ) -> Result<Vec<TransactionMetadata<Block>>, String> {
+    //     match self.db.get(crate::columns::TRANSACTION_MAPPING, &starknet_transaction_hash.encode()) {
+    //         Some(raw) => Ok(Vec::<TransactionMetadata<Block>>::decode(&mut &raw[..]).map_err(|e|
+    // format!("{:?}", e))?),         None => Ok(Vec::new()),
+    //     }
+    // }
 
+    /// Register that a Substrate block has been seen, without it containing a Starknet one
     pub fn write_none(&self, block_hash: Block::Hash) -> Result<(), String> {
         let _lock = self.write_lock.lock();
 
@@ -67,6 +77,7 @@ impl<Block: BlockT> MappingDb<Block> {
         Ok(())
     }
 
+    /// Register that a Substate block has been seen and map it to the Statknet block it contains
     pub fn write_hashes(&self, commitment: MappingCommitment<Block>) -> Result<(), String> {
         let _lock = self.write_lock.lock();
