+++ crates/client/db/src/meta_db.rs
@@ -8,12 +8,17 @@ use sp_runtime::traits::Block as BlockT;
 
 use crate::DbHash;
 
+/// Allow interaction with the meta db
+///
+/// The meta db store the tips of the synced chain.
+/// In case of forks, there can be multiple tips.
 pub struct MetaDb<Block: BlockT> {
     pub(crate) db: Arc<dyn Database<DbHash>>,
     pub(crate) _marker: PhantomData<Block>,
 }
 
 impl<Block: BlockT> MetaDb<Block> {
+    /// Retrieve the currrent tips of the synced chain
     pub fn current_syncing_tips(&self) -> Result<Vec<Block::Hash>, String> {
         match self.db.get(crate::columns::META, crate::static_keys::CURRENT_SYNCING_TIPS) {
             Some(raw) => Ok(Vec::<Block::Hash>::decode(&mut &raw[..]).map_err(|e| format!("{:?}", e))?),
@@ -21,6 +26,7 @@ impl<Block: BlockT> MetaDb<Block> {
         }
     }
 
+    /// Store the current tips of the synced chain
     pub fn write_current_syncing_tips(&self, tips: Vec<Block::Hash>) -> Result<(), String> {
         let mut transaction = sp_database::Transaction::new();
